chore: move IO logic into separate module
